Bug 1581024 - fixup pointer wrap check to prevent it from being optimized out. r=jcj
Differential Revision: https://phabricator.services.mozilla.com/D47013

--HG--
extra : moz-landing-system : lando

+++ cmd/lib/derprint.c
@@ -509,7 +509,8 @@ prettyPrintItem(FILE *out, const unsigned char *data, const unsigned char *end,
         /*
          * Just quit now if slen more bytes puts us off the end.
          */
-        if ((data + slen) < data || (data + slen) > end) {
+        if (((unsigned long)data + slen) < (unsigned long)data ||
+            ((unsigned long)data + slen) > (unsigned long)end) {
             PORT_SetError(SEC_ERROR_BAD_DER);
             return -1;
         }
